feat(outputs): allow to specify context for output handlers via Tokens
Add `IoEventContextToken` and `IoEventContextProviderToken` that will be used as a context when
output handlers are invoked.

Now `IoService` is part of the public API and you may use it to bind inputs/outputs imperatively
anywhere from your code. To provide it use `IoServiceProvider` to remain BC compatible in the future.

Also most of the codebase was refactored to replace `any` with more typesafe `unknown` and one bug
in dynamic directive ref was fixed when host component was typed as a class instead of an instance.

DEPRECATIONS:
- `EventArgumentToken`: please use `IoEventArgumentToken`